+++ module/data/_module.mjs
@@ -1,3 +1,8 @@
+export * as fields from "./fields/_module.mjs";
+
+export * as activeEffect from "./active-effect/_module.mjs";
+export * as actor from "./actor/_module.mjs";
+export * as item from "./item/_module.mjs";
 export * as regionBehavior from "./region-behaviors/_module.mjs";
 
 export { default as TerrainData4e } from "./terrain-data.mjs";
