+++ module/data/active-effect/_module.mjs
@@ -1,0 +1,1 @@
+export { default as ActiveEffectData } from "./active-effect.mjs";
