+++ module/data/active-effect/active-effect.mjs
