+++ module/data/actor/_module.mjs
@@ -1,0 +1,6 @@
+export * as fields from "./fields/_module.mjs";
+export * as templates from "./templates/_module.mjs";
+
+export { default as CharacterData } from "./character.mjs";
+export { default as HazardData } from "./hazard.mjs";
+export { default as NPCData } from "./npc.mjs";
