+++ module/data/actor/fields/_module.mjs
@@ -1,0 +1,3 @@
+export { default as BonusField } from "./bonus-field.mjs";
+export { default as Dnd4eBonusesField } from "./dnd4e-bonuses-field.mjs";
+export { default as SimpleTraitField } from "./simple-trait-field.mjs";
