+++ module/data/actor/npc.mjs
@@ -1,12 +1,6 @@
-import BonusField from "./fields/bonus-field.js";
-import Dnd4eBonusesField from "./fields/dnd4e-bonuses-field.js";
-import MappingField from "../fields/mapping-field.js";
-import AttributesField from "./templates/attributes.js";
-import CombatantTemplate from "./templates/combatant.js";
-import CreatureTemplate from "./templates/creature.js";
-import DetailsField from "./templates/details.js";
-import SpeedTemplate from "./templates/speed.js";
-import WealthTemplate from "./templates/wealth.js";
+import { BonusField, Dnd4eBonusesField } from "./fields/_module.mjs";
+import MappingField from "../fields/mapping-field.mjs";
+import { AttributesField, CombatantTemplate, CreatureTemplate, DetailsField, SpeedTemplate, WealthTemplate } from "./templates/_module.mjs";
 
 const { BooleanField, NumberField, StringField, SchemaField } = foundry.data.fields;
