+++ module/data/actor/templates/_module.mjs
@@ -1,0 +1,7 @@
+export { default as AttributesField } from "./attributes.mjs";
+export { default as CombatantTemplate } from "./combatant.mjs";
+export { default as CreatureTemplate } from "./creature.mjs";
+export { default as DetailsField } from "./details.mjs";
+export { default as SensesTemplate } from "./senses.mjs";
+export { default as SpeedTemplate } from "./speed.mjs";
+export { default as WealthTemplate } from "./wealth.mjs";
