+++ module/data/fields/_module.mjs
@@ -1,0 +1,2 @@
+export { default as FormulaField } from "./formula-field.mjs";
+export { default as MappingField } from "./mapping-field.mjs";
