+++ module/data/item/_module.mjs
@@ -1,0 +1,12 @@
+export * as fields from "./fields/_module.mjs";
+export * as templates from "./templates/_module.mjs";
+
+export { default as BackpackData } from "./backpack.mjs";
+export { default as ConsumableData } from "./consumable.mjs";
+export { default as EquipmentData } from "./equipment.mjs";
+export { default as FeatureData } from "./feature.mjs";
+export { default as LootData } from "./loot.mjs";
+export { default as PowerData } from "./power.mjs";
+export { default as RitualData } from "./ritual.mjs";
+export { default as ToolData } from "./tool.mjs";
+export { default as WeaponData } from "./weapon.mjs";
