+++ module/data/item/_utils.mjs
@@ -1,4 +1,4 @@
-/*
+/**
 * Utility function to migrate a damage part from a string array to an object.
 * @param {Array} part       The damage part to migrate
 * @return {object} part     The migrated part
