+++ module/data/item/consumable.mjs
@@ -1,11 +1,6 @@
-import FormulaField from "../fields/formula-field.js";
-import MappingField from "../fields/mapping-field.js";
-import ActivatedEffectTemplate from "./templates/activated-effect.js";
-import AttackAndDamageTemplate from "./templates/attack-damage.js";
-import ItemDescriptionTemplate from "./templates/item-description.js";
-import ItemMacroTemplate from "./templates/item-macro.js";
-import PhysicalItemTemplate from "./templates/physical-item.js";
-import { processPart } from "./_utils.js";
+import { FormulaField, MappingField } from "../fields/_module.mjs";
+import { ActivatedEffectTemplate, AttackAndDamageTemplate, ItemDescriptionTemplate, ItemMacroTemplate, PhysicalItemTemplate } from "./templates/_module.mjs";
+import { processPart } from "./_utils.mjs";
 
 const { BooleanField, NumberField, StringField, SchemaField } = foundry.data.fields;
