+++ module/data/item/fields/_module.mjs
@@ -1,0 +1,1 @@
+export { default as DamagePartsField } from "./damage-parts-field.mjs";
