+++ module/data/item/templates/_module.mjs
@@ -1,0 +1,5 @@
+export { default as ActivatedEffectTemplate } from "./activated-effect.mjs";
+export { default as AttackAndDamageTemplate } from "./attack-damage.mjs";
+export { default as ItemDescriptionTemplate } from "./item-description.mjs";
+export { default as ItemMacroTemplate } from "./item-macro.mjs";
+export { default as PhysicalItemTemplate } from "./physical-item.mjs";
